gcp/requirements.txt
Flask==3.0.3
tensorflow==2.13.0
google-cloud-storage==2.19.0
Pillow==10.0.0
numpy==1.24.3